% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{center.preds}
\alias{center.preds}
\title{Centers the predicted values for 1-d ICE and PDP plots, 2-d PDP plots}
\usage{
center.preds(object, plot.type)
}
\arguments{
\item{object}{The Plotter object to use}

\item{plot.type}{The type of plot which the user wants to center}
}
\value{
centered dataframe/matrix of values for the given plot
}
\description{
Given the specified 'center.at' values of the plotter object, this
             function centers the speccified type of plot.
}
\details{
center.preds
}
