% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distiller.R
\name{distill}
\alias{distill}
\title{Build surrogate model from distilled model}
\usage{
distill(
  object,
  center.mean = T,
  snap.grid = T,
  fit.train = T,
  params.glmnet = list()
)
}
\arguments{
\item{object}{The Interpreter object}

\item{center.mean}{Boolean value that determines whether to center each column
of predictions by their respective means. Default is TRUE}

\item{snap.grid}{Boolean function that determines whether the model recalculates
each value predicted or uses an approximation from previous
calculations. Default is TRUE.}

\item{fit.train}{Fit to training data or fit to expand.grid. If true, we fit to
the training data. Default is TRUE.}

\item{params.glmnet}{Optional list of parameters to pass to glmnet while fitting
PDP curves to resemble the original predictions. By specifying
parameters, one can do lasso or ridge regression.}
}
\value{
A surrogate class object that can be used for predictions
}
\description{
Builds a surrogate model from the PDP functions
}
