% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{localSurrogate}
\alias{localSurrogate}
\title{Given a interpreter object with at least one pair of features,
  create a small surrogate model for the two features using the PDP function
  as the output and the two features as the independent variables.}
\usage{
localSurrogate(object, interact = FALSE, params.forestry = list())
}
\arguments{
\item{object}{Interpreter object to make plots + surrogate for.}

\item{interact}{An indicator specifying if the surrogate model should also be
given the interaction terms to create the surrogate models with.
Default is FALSE.}

\item{params.forestry}{Optional list of parameters to pass to the surrogate
model. Defaults to the standard Rforestry parameters with ntree = 1
and maxDepth = 2.}
}
\description{
Plots and returns a Rforestry object with a single tree explaining
  the PDP surface.
}
