% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{pdpRisk}
\alias{pdpRisk}
\title{Given an interpreter object with choice of PDP ranking methodology
       (default: 'Variance'), produce 'risk' scores by feature. Optionally,
       permits a new observation to weight the PDP function and rankings.}
\usage{
pdpRisk(object, rank.method = "Variance", new.obs = TRUE)
}
\arguments{
\item{rank.method}{A string to select which PDP ranking methodology. Should be one of
c("Variance", "FO.Derivative"). When set to "Variance" the PDP functions are ranked by variance
of the function over the range of grid.points and when set to "FO.Derivative" the PDP functions
are ranked by empirical first order derivative.}

\item{new.obs}{A data frame with equivalent name/features of the data that trained the
predictor parameter, holding information/statistics of an observation of interest.
A non-empty data frame re-weights the PDP function by impact of similar observations
in the training data on the response variable.}
}
\value{
A list of risk scores by feature.
}
\description{
Returns a list of 'risk' scores corresponding to each feature.
}
