% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot-Interpreter}
\alias{plot-Interpreter}
\alias{plot.Interpreter}
\title{PLotting method for Interpretor model}
\usage{
\method{plot}{Interpreter}(x, method = "pdp+ice", features = NULL, features.2d = NULL, ...)
}
\arguments{
\item{x}{Interpreter object to make plots for.}

\item{method}{A method for the plot. Must be one of "ice", "pdp+ice", or "pdp"}

\item{features}{1-D PDP/ICE curves to plot.}

\item{features.2d}{2-D PDP curves to plot.}

\item{...}{Additional arguments}
}
\value{
A list of plots with 1-d features and 2-d features. For 2-d features with
        one continuous and one categorical feature, the plot is a linear plot of the
        continuous feature with group colors representing the categorical feature.
        For two continuous features, the plot is a heatmap with the shade representing
        the value of the outcome.
}
\description{
Plots either the PDP plots or ICE plots
}
