% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot-Plotter}
\alias{plot-Plotter}
\alias{plot.Plotter}
\title{PLotting method for Interpretor model}
\usage{
\method{plot}{Plotter}(x, method = "pdp+ice", ...)
}
\arguments{
\item{x}{Plotter object to make plots for.}

\item{method}{A method for the plot. Must be one of "ice", "pdp+ice", or "pdp"}

\item{...}{Additional arguments}
}
\value{
A list of plots with 1-d features and 2-d features. For 2-d features with
        one continuous and one categorical feature, the plot is a linear plot of the
        continuous feature with group colors representing the categorical feature.
        For two continuous features, the plot is a heatmap with the shade representing
        the value of the outcome.
}
\description{
Plots either the PDP plots or ICE plots
}
