% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{plot.Interpretor}
\alias{plot.Interpretor}
\title{PLotting method for Interpretor model}
\usage{
\method{plot}{Interpretor}(object, method = "pdp+ice")
}
\arguments{
\item{object}{Interpretor object to make plots for.}

\item{method}{The type of plot to make. Can be one of `ice`,`pdp`,`pdp+ice`.}
}
\description{
Plots either the PDP plots or ICE plots
}
