% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict-Predictor}
\alias{predict-Predictor}
\alias{predict.Predictor}
\title{Predict method for Predictor class}
\usage{
\method{predict}{Predictor}(object, newdata, ...)
}
\arguments{
\item{object}{The Predictor object to use to make predictions.}

\item{newdata}{The data frame to use for the independent features in the prediction.}

\item{...}{Additional arguments that are passed to the model predict function.
For instance, these can be different aggregation options (aggregation = "oob")
that are accepted by the prediction function of the model.}
}
\value{
A data frame with a single column containing the predictions for each
 row of the newdata data frame.
}
\description{
Gives a single column of predictions from a model that
 is wrapped by the Predictor object
}
