% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surrogate.R
\name{predict-Surrogate}
\alias{predict-Surrogate}
\alias{predict.Surrogate}
\title{Prediction method for the distilled surrogate model}
\usage{
\method{predict}{Surrogate}(object, newdata, ...)
}
\arguments{
\item{object}{A surrogate object distilled from the interpreter}

\item{newdata}{The dataframe to use for the predictions}

\item{...}{Additional parameters to pass to predict}
}
\value{
A one-column dataframe of the surrogate model's predictions
}
\description{
Predicts outputs given new data
}
