% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{predict.Predictor}
\alias{predict.Predictor}
\title{Predict method for Predictor class}
\usage{
\method{predict}{Predictor}(object, newdata)
}
\arguments{
\item{object}{Predictor object to use.}

\item{newdata}{The dataframe to use for the predictions.}
}
\description{
Gives predictions for a predictor object
}
