% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{predict_ALE}
\alias{predict_ALE}
\title{Prediction function for the ALE plots}
\usage{
predict_ALE(x, feature, training_data, save = TRUE)
}
\arguments{
\item{x}{An interpreter object}

\item{feature}{The feature to build the ALE for (must be continuous)}

\item{training_data}{The training data to use in order to build the ALE}

\item{save}{Boolean to save the ALE predictions}
}
\value{
A tibble that contains the ALE predictions for the given values
}
\description{
Prediction function for the ALE plots
}
