% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{predict_grid.Interpretor}
\alias{predict_grid.Interpretor}
\title{# prediction function for grid points}
\usage{
predict_grid.Interpretor(object)
}
\arguments{
\item{object}{The Interpretor object to use.}
}
\description{
Gives predictions at each point on the grid.
}
