% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{predict_grid.Plotter}
\alias{predict_grid.Plotter}
\title{Prediction Function for ICE Plots}
\usage{
predict_grid.Plotter(object)
}
\arguments{
\item{object}{The Plotter object to use.

Needed fixes: set predictions as a one time calculation,
add in second variable interpretations,
add grouping by another variable}
}
\description{
Gives predictions at each point on the grid.
}
