% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{predict_ICE.Plotter}
\alias{predict_ICE.Plotter}
\title{Prediction Function for ICE Plots}
\usage{
predict_ICE.Plotter(object, save = TRUE)
}
\arguments{
\item{object}{The Plotter object to use.}

\item{save}{A binary variable to determine whether the calculations should be
saved in the plotter object}
}
\value{
A list of dataframes for each feature. In each dataframe, the first
        column contains the grid values for the feature, and each subsequent
        column has a single observation with the modified feature set to that
        row's grid point.

Needed fixes: add in second variable interpretations,
add grouping by another variable
}
\description{
Gives predictions at each point on the grid.
}
