% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{predict_PDP.1D.Plotter}
\alias{predict_PDP.1D.Plotter}
\title{Prediction Function for PDP Plots}
\usage{
predict_PDP.1D.Plotter(object, save = TRUE)
}
\arguments{
\item{object}{The Interpreter object to use.}

\item{save}{A binary variable to determine whether the calculations should be
saved in the interpreter object}
}
\value{
A list of dataframes with the grid points and PDP prediction curves
        for each feature in object
}
\description{
Gives prediction curve for all specified features in the
             plotter object
}
\details{
predict_PDP.1D.Plotter
}
