% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{predict_PDP.2D.Plotter}
\alias{predict_PDP.2D.Plotter}
\title{Two Dimensional Prediction Curve for PDP Plots}
\usage{
predict_PDP.2D.Plotter(object, feat.2d, save = TRUE)
}
\arguments{
\item{object}{The Interpreter object to use.}

\item{feat.2d}{A 2-column dataframe or matrix that gives the first variable in
in the first column, and the second variable in the next. This should
the number of rows as the number of 2-D PDPs one would like}

\item{save}{A binary variable to determine whether the calculations should be
saved in the interpreter object}
}
\value{
A list of dataframes for each pair of features.2d
}
\description{
Gives prediction surface for all specified feature pairs in the
             interpreter object (features.2d)
}
\details{
predict_PDP.2D.Plotter
}
