% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{Predictor}
\alias{Predictor}
\title{The class that wraps a machine learning model in order to provide a
 standardized method for predictions for different models.}
\description{
Predictor class description
}
\details{
Predictor Object
}
\note{
A wrapper to pass an ML algorithm (rpart, etc.) through the
interpretability functions, the data used to create the algorithm

The necessary variables are model, data, y. The other variables are
optional, and depend on the use cases. Type should be used only when
a prediction function is NOT specified.

The outputs of the algorithm must be the values if it is regression, or
probabilities if classification. For classification problems with more than
two categories, the output comes out as vectors of probabilities for each
category. Because this is for ML interpretability, other types of
predictions (ex: predictions that spit out the factor) are not allowed.
prediction method must be constructed, with optional arguments of
type and task
}
\examples{
data <- MASS::Boston
set.seed(491)
test_ind <- sample(1:nrow(data), nrow(data)\%/\%5)
train_reg <- data[-test_ind,]
test_reg <- data[test_ind,]


data_class <- MASS::crabs
test_ind <- sample(1:nrow(data_class), nrow(data_class)\%/\%5)
train_class <- data_class[-test_ind,]
test_class <- data_class[test_ind,]


linreg <- lm(crim ~., data=train_reg)
linreg_predictor <- Predictor$new(model = linreg, data = train_reg, y="crim",
                                  predict = predict, task = "regression")
actual <- predict(linreg, test_reg)
method <- linreg_predictor$predict(test_reg)
sum(actual != method) # all the same values

linreg_predictor$print()
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The training data for building the model}

\item{\code{model}}{The object model}

\item{\code{task}}{The prediction task the model performs (i.e. classification or regression)}

\item{\code{class}}{The class for which we get predictions. We specify this to get the predictions
(such as probabilites) for an observation being in a specific class (e.g. Male or Female).
This parameter is specifically for classification predictions.}

\item{\code{prediction.function}}{The prediction function used by the model}

\item{\code{batch.size}}{The number of observations used for future predictions. This is at most the number of
observations in the training data.}

\item{\code{y}}{The name of the outcome feature in `data`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Predictor$new()}}
\item \href{#method-clone}{\code{Predictor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$new(
  model = NULL,
  data = NULL,
  predict.func = NULL,
  y = NULL,
  task = NULL,
  class = NULL,
  type = NULL,
  batch.size = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A trained model.}

\item{\code{data}}{A dataframe containing the training data.}

\item{\code{predict.func}}{A function to run predictions for the model.}

\item{\code{y}}{The name of the outcome variable in data.}

\item{\code{task}}{The prediction class, either `classification` or `regression`}

\item{\code{class}}{The class of predictions done.}

\item{\code{type}}{The type of predictions done (i.e. 'response' for predicted probabliities for classification).
This feature is only be used if no specific prediction function is specified.}

\item{\code{batch.size}}{The size of the batch used to create pdp grids}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `Predictor` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
