% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{print-Predictor}
\alias{print-Predictor}
\alias{print.Predictor}
\title{Printing method for Predictor class}
\usage{
\method{print}{Predictor}(x, ...)
}
\arguments{
\item{x}{The Predictor to print}

\item{...}{Additional arguments}
}
\description{
Gives the task of the given predictor x
}
