% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{print.Predictor}
\alias{print.Predictor}
\title{Printing method for Predictor class}
\usage{
\method{print}{Predictor}(object)
}
\arguments{
\item{object}{The Predictor to print}
}
\description{
Gives information for a predictor object
}
