% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{set.data.points}
\alias{set.data.points}
\title{Modify the Training Data Used for Interpretability}
\usage{
set.data.points(object, data.points)
}
\arguments{
\item{object}{The Interpreter object}

\item{data.points}{The training data indices used to set which observations
are used for interpretability}
}
\description{
Sets the data points used for interpretation
}
