% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{set.grid.points}
\alias{set.grid.points}
\title{Sets the center value used for plotting}
\usage{
set.grid.points(object, feature, values)
}
\arguments{
\item{object}{The Plotter class to recenter}

\item{feature}{The feature to center}

\item{values}{The set of new values to be used as the grid points for the selected feature}
}
\description{
Method for setting center value for a specific feature
}
