% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{set.method}
\alias{set.method}
\title{Modify the Method Used for Interpretability}
\usage{
set.method(object, method)
}
\arguments{
\item{object}{The Interpreter object to change the method attribute in}

\item{method}{The new method to use in the Interpreter object. Must be one
of "pdp" or "ale"}
}
\description{
Sets a new method used for interpretability in the Interpreter
  object. This can be different from the previous type ("ale" or "pdp")
}
