% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distiller.R
\name{build.grid}
\alias{build.grid}
\title{Build grid used for weights in distilled surrogate model}
\usage{
build.grid(object, feat.ind = 1:length(object$features))
}
\arguments{
\item{object}{The Interpreter object}

\item{feat.ind}{The indices of the features in the Interpreter's features that we want
to include as PDP functions in the distilled model.}
}
\value{
A dataframe used to find weights in regression (one-hot encoding for
        categorical features)
}
\description{
A dataframe storing the true predictions and the PDP predictions
}
\note{
This function is mainly used as a subroutine for the distill function. We include this
as a public function to allow users to create their own weights and surrogate functions
outside of our implemented method.
}
