% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{Interpreter}
\alias{Interpreter}
\title{Interpreter}
\description{
Interpreter

Interpreter

set.gridpoints.Interpreter

set.center.Interpreter

predict_grid.Interpreter

plot.Interpreter
}
\note{
PDP and ICE Implementation
similar methods
Needed fixes: set predictions as a one time calculation, allow for specific
points instead of number of points, add in second variable interpretations,
add grouping by another variable
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Interpreter$new()}}
\item \href{#method-set.gridpoints}{\code{Interpreter$set.gridpoints()}}
\item \href{#method-set.center}{\code{Interpreter$set.center()}}
\item \href{#method-predict_grid}{\code{Interpreter$predict_grid()}}
\item \href{#method-plot}{\code{Interpreter$plot()}}
\item \href{#method-clone}{\code{Interpreter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$new(
  predictor,
  feature,
  center.at = NULL,
  grid.size = 20,
  grid.points = NULL,
  group = NULL,
  samples = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{The trained predictor object for the model that we want to
interpret.}

\item{\code{feature}}{The name of the feature that we use for the interpretability model.}

\item{\code{center.at}}{The location we should use to center the predictions.}

\item{\code{grid.size}}{The grid size used for the PDP.}

\item{\code{grid.points}}{Number of grid points used in the PDP.}

\item{\code{group}}{A group variable to classify the observations and give more
granular predictions.}

\item{\code{samples}}{The samples used for the PDP/ALE.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An `Interpreter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.gridpoints"></a>}}
\if{latex}{\out{\hypertarget{method-set.gridpoints}{}}}
\subsection{Method \code{set.gridpoints()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$set.gridpoints(grid.points)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{grid.points}}{The grid points to be used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set.center"></a>}}
\if{latex}{\out{\hypertarget{method-set.center}{}}}
\subsection{Method \code{set.center()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$set.center(center = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{center}}{The value to center the predictions at}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict_grid"></a>}}
\if{latex}{\out{\hypertarget{method-predict_grid}{}}}
\subsection{Method \code{predict_grid()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$predict_grid()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\if{latex}{\out{\hypertarget{method-plot}{}}}
\subsection{Method \code{plot()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$plot(method = "pdp+ice")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{method}}{The type of plot to make. Can be one of `ice`,`pdp`,`pdp+ice`.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
