% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{Interpreter}
\alias{Interpreter}
\title{The class for different interpretability methods.}
\description{
Interpreter class description
}
\note{
PDP and ICE Implementation
}
\examples{
Needs to be rewritten
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{predictor}}{The predictor object to use as a standardised wrapper for the model}

\item{\code{data.points}}{The indices of the data points used for the PDP/ALE. This
overwrites the "samples" initialization or the batch size variable from the predictor class.}

\item{\code{functions.1d}}{Functions giving average value across data.points with a given value and feature}

\item{\code{functions.2d}}{Functions giving 2D pdp functions}

\item{\code{method}}{The chosen interpretability method for the black-box model.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Interpreter$new()}}
\item \href{#method-clone}{\code{Interpreter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$new(
  predictor = NULL,
  samples = NULL,
  data.points = NULL,
  method = "pdp"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{predictor}}{The trained predictor object for the model that we want to
interpret.}

\item{\code{samples}}{The samples used for the PDP/ALE. Overwrites the batch size
parameter from the predictor object.}

\item{\code{data.points}}{The indices of the data points used for the PDP/ALE. This
overwrites the "samples" parameter or the batch size from the predictor.}

\item{\code{method}}{The chosen interpretability method for the black-box model.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
An `Interpreter` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Interpreter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
