% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{plot-Interpreter}
\alias{plot-Interpreter}
\alias{plot.Interpreter}
\title{Plotting method for Interpretor model}
\usage{
\method{plot}{Interpreter}(x, method = "pdp+ice", features = NULL, features.2d = NULL, ...)
}
\arguments{
\item{x}{Interpreter object to generate plots from}

\item{method}{The type of plot that we want to generate. Must be one of "ice",
"pdp+ice", "pdp", or "ale"}

\item{features}{1-D features that we want to produce plots for.}

\item{features.2d}{2-D features that we want to produce plots for arguments}
}
\value{
A list of plots with 1-d features and 2-d features. For 2-d features with
        one continuous and one categorical feature, the plot is a linear plot of the
        continuous feature with group colors representing the categorical feature.
        For two continuous features, the plot is a heatmap with the shade representing
        the value of the outcome.
}
\description{
Plots the PDP, ALE, or ICE plots for an Interpreter object
}
