% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{Plotter}
\alias{Plotter}
\title{The class for interpretability plots}
\description{
Plotter class description
}
\examples{
Needs to be rewritten
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{interpreter}}{The interpreter object to use to create the plots.}

\item{\code{features}}{The set of features for which to create plots for.}

\item{\code{features.2d}}{Pairs of features to create 2-d plots for.}

\item{\code{center.at}}{The value to center the plots at}

\item{\code{grid.points}}{The grid points to use for the predictions.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Plotter$new()}}
\item \href{#method-clone}{\code{Plotter$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Initialize a plotter object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plotter$new(
  interpreter = NULL,
  features = NULL,
  features.2d = NULL,
  grid.size = 30
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{interpreter}}{Interpreter object to make plots for.}

\item{\code{features}}{The features to be used for the interpretor plot}

\item{\code{features.2d}}{A matrix or data frame containing two columns of
feature names to give pairs of features for the 2-d plots.}

\item{\code{grid.size}}{The number of grid points to create for the pdp
functions / plots.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Plotter$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
