% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{predict_PDP.2D.Plotter}
\alias{predict_PDP.2D.Plotter}
\title{Two Dimensional Prediction Curve for PDP Plots}
\usage{
predict_PDP.2D.Plotter(object, save = TRUE)
}
\arguments{
\item{object}{The Interpreter object to use.}

\item{save}{A binary variable to determine whether the calculations should be
saved in the interpreter object}
}
\value{
A list of dataframes for each pair of features.2d
}
\description{
Gives prediction surface for all specified feature pairs in the
             interpreter object (features.2d)
}
\details{
predict_PDP.2D.Plotter
}
