% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{Predictor}
\alias{Predictor}
\title{Predictor}
\description{
Predictor Object
}
\note{
A wrapper to pass through an ML algorithm (rpart, etc.) through the
interpretability functions, the data used to create the algorithm

The necessary variables are model, data, y. The other variables are
optional, and depend on the use cases. Type should be used only when
a prediction function is NOT specified.

The outputs of the algorithm must be the values if it is regression, or
probabilities if classification. For classification problems with more than
two categories, the output comes out as vectors of probabilities for each
category. Because this is for ML interpretability, other types of
predictions (ex: predictions that spit out the factor) are not allowed.
prediction method must be constructed, with optional arguments of
type and task
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Predictor$new()}}
\item \href{#method-predict}{\code{Predictor$predict()}}
\item \href{#method-print}{\code{Predictor$print()}}
\item \href{#method-clone}{\code{Predictor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$new(
  model = NULL,
  data = NULL,
  predict.func = NULL,
  y = NULL,
  task = NULL,
  class = NULL,
  type = NULL,
  batch.size = 1000
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A trained model.}

\item{\code{data}}{A dataframe containing the training data.}

\item{\code{predict.func}}{A function to run predictions for the model.}

\item{\code{y}}{The name of the outcome variable in data.}

\item{\code{task}}{The prediction class, either `classification` or `regression`}

\item{\code{class}}{The name of the outcome variable in data.}

\item{\code{type}}{The type of regression.}

\item{\code{batch.size}}{The size of the batch used to create pdp grids}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `Predictor` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-predict"></a>}}
\if{latex}{\out{\hypertarget{method-predict}{}}}
\subsection{Method \code{predict()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$predict(newdata)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{newdata}}{The data frame to run predictions on.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
