% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{Predictor}
\alias{Predictor}
\title{Predictor class description}
\description{
The class that wraps a machine learning model in order to provide a
 standardized method for predictions for different models.
}
\details{
Predictor Object
}
\note{
A wrapper to pass an ML algorithm (rpart, etc.) through the
interpretability functions, the data used to create the algorithm

The necessary variables are model, data, y. The other variables are
optional, and depend on the use cases. Type should be used only when
a prediction function is NOT specified.

The outputs of the algorithm must be the values if it is regression, or
probabilities if classification. For classification problems with more than
two categories, the output comes out as vectors of probabilities for the
specified "class" category. Because this is for ML interpretability,
other types of predictions (ex: predictions that spit out the factor) are not allowed.
prediction method must be constructed, with optional argument of type
}
\examples{
library(interpret)
library(Rforestry)
set.seed(491)
data <- MASS::crabs

levels(data$sex) <- list(Male = "M", Female = "F")
levels(data$sp) <- list(Orange = "O", Blue = "B")
colnames(data) <- c("Species","Sex","Index","Frontal Lobe",
"Rear Width", "Carapace Length","Carapace Width","Body Depth")

test_ind <- sample(1:nrow(data), nrow(data)\%/\%5)
train_reg <- data[-test_ind,]
test_reg <- data[test_ind,]


forest <- forestry(x=train_reg[,-which(names(train_reg)=="Carapace Width")],
y=train_reg[,which(names(train_reg)=="Carapace Width")])

forest_predictor <- Predictor$new(model = forest, data=train_reg,
y="Carapace Width", task = "regression")


}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{The training data for building the model}

\item{\code{model}}{The object model}

\item{\code{task}}{The prediction task the model performs (i.e. classification or regression)}

\item{\code{class}}{The class for which we get predictions. We specify this to get the predictions
(such as probabilites) for an observation being in a specific class (e.g. Male or Female).
This parameter is necessary for classification predictions with more than a single vector
of predictions.}

\item{\code{prediction.function}}{The prediction function used by the model}

\item{\code{y}}{The name of the outcome feature in `data`.}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Predictor$new()}}
\item \href{#method-clone}{\code{Predictor$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$new(
  model = NULL,
  data = NULL,
  predict.func = NULL,
  y = NULL,
  task = NULL,
  class = NULL,
  type = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{model}}{A trained model.}

\item{\code{data}}{A dataframe containing the training data.}

\item{\code{predict.func}}{A function to run predictions for the model.}

\item{\code{y}}{The name of the outcome variable in data.}

\item{\code{task}}{The prediction class, either `classification` or `regression`}

\item{\code{class}}{The class of outcomes we want to see predictions for.}

\item{\code{type}}{The type of predictions done (i.e. 'response' for predicted probabliities for classification).
This feature should only be used if no predict.funcc is specified.}
}
\if{html}{\out{</div>}}
}
\subsection{Returns}{
A `Predictor` object.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Predictor$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
