% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictor.R
\name{print-Predictor}
\alias{print-Predictor}
\alias{print.Predictor}
\title{The Printing method for Predictor class}
\usage{
\method{print}{Predictor}(x, ...)
}
\arguments{
\item{x}{The Predictor object to print}

\item{...}{Additional arguments passed to the print function.}
}
\description{
Prints the task of an instance of the Predictor class.
}
