% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{set.center.at}
\alias{set.center.at}
\title{Sets a new center in the PDP and ICE plots made by an Interpreter}
\usage{
set.center.at(object, feature, value)
}
\arguments{
\item{object}{The Interpreter class that we want to recenter the plots of.}

\item{feature}{The name of the feature to set grid points for.}

\item{value}{The new value to use for the plots of the specified feature to be centered at.
Must match the type of the feature (a factor level or continuous value in the range
of the specified feature).}
}
\description{
Method for setting center value for a specific feature
}
\note{
Unlike the grid predictions, the center.at values do not modify any of the
previous saved calculations. Therefore, it does not change or remove any of the
previously calculated, saved data. These center values are simply for the plots
made by the interpreter object, rather than the distilled model.
}
