% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{set.center.at}
\alias{set.center.at}
\title{Sets the center value used for plotting}
\usage{
set.center.at(object, feature, value)
}
\arguments{
\item{object}{The Plotter class to recenter}

\item{feature}{The feature to center}

\item{value}{The new center value to use for the feature.}
}
\description{
Method for setting center value for a specific feature
}
\note{
Unlike the grid predictions, the center.at values do not modify any of the
previous saved calculations. Therefore, it does not change or remove any of the
previously calculated, saved data.
}
