% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{set.data.points}
\alias{set.data.points}
\title{Sets the data points used for interpretation}
\usage{
set.data.points(object, data.points)
}
\arguments{
\item{object}{The Interpreter to print}

\item{data.points}{The new data points to set for the object}
}
\description{
Gives information for a predictor object
}
