% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotter.R
\name{set.grid.points}
\alias{set.grid.points}
\title{Sets the center value used for plotting}
\usage{
set.grid.points(object, feature, values)
}
\arguments{
\item{object}{The Interpreter class to recenter}

\item{feature}{The feature to center}

\item{values}{The set of new values to be used as the grid points for the selected feature}
}
\description{
Method for setting center value for a specific feature
}
\note{
Because the grid points determine what calculations are performed, changing the grid points
will remove any of the previously calculated values in the 'Interpreter' object. For any 1-d ICE
or PDP plot, it will remove the previous calculations for the given feature. For any 2-d PDP
calcuations, it will remove plots that include the given feature as any of its features.
}
