% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interpret.R
\name{set.method}
\alias{set.method}
\title{Modify the Method Used for Interpretability}
\usage{
set.method(object, method)
}
\arguments{
\item{object}{The Interpreter object}

\item{method}{The new method to generate the interpretability functions}
}
\description{
Sets a new method used for interpertability
}
